requirements.txt
requests
google-generativeai>=0.4.0
python-dotenv
supabase